refactor: factor out decision module
